backend/requirements.txt
# Core backend
fastapi
uvicorn[standard]
gunicorn

# Database
sqlalchemy>=2.0.47      # Required for Python 3.14 compatibility
psycopg[binary]         # Modern PostgreSQL driver

# Templates & forms
jinja2
python-multipart

# Authentication & sessions
authlib
httpx
itsdangerous

# Password hashing
passlib[bcrypt]
bcrypt

# Environment variables
python-dotenv